+++ src/main/java/com/demo/classroom/Utility/Exception/UserAlreadyExistsException.java
