--- src/main/java/com/demo/classroom/Validation/PasswordValidator.java
@@ -1,22 +1,0 @@
-package com.demo.classroom.Validation;
-
-import jakarta.validation.ConstraintValidator;
-import jakarta.validation.ConstraintValidatorContext;
-
-public class PasswordValidator implements ConstraintValidator<ValidPassword, String> {
-
-    private static final String PASSWORD_PATTERN =
-            "^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=!])(?=\\S+$).{8,}$";
-
-    @Override
-    public void initialize(ValidPassword constraintAnnotation) {
-    }
-
-    @Override
-    public boolean isValid(String password, ConstraintValidatorContext context) {
-        if (password == null) {
-            return false;
-        }
-        return password.matches(PASSWORD_PATTERN);
-    }
-}
